requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.2
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.1
pytz==2021.3
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
tablib==3.2.0
tzdata==2021.5
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0
